+++ aioxmpp/muc/service.py
@@ -806,7 +806,7 @@ class Room:
         )
         msg.subject.update(subject)
 
-        return self.service.client.stream.enqueue_stanza(msg)
+        return self.service.client.stream.enqueue(msg)
 
     def leave(self):
         """
@@ -826,7 +826,7 @@ class Room:
             type_=aioxmpp.structs.PresenceType.UNAVAILABLE,
             to=self._mucjid
         )
-        self.service.client.stream.enqueue_stanza(presence)
+        self.service.client.stream.enqueue(presence)
 
     @asyncio.coroutine
     def leave_and_wait(self):
@@ -901,7 +901,7 @@ class Room:
             message.body.update(body_or_stanza)
 
         tracker = aioxmpp.tracking.MessageTracker()
-        token = self.service.client.stream.enqueue_stanza(
+        token = self.service.client.stream.enqueue(
             message,
             on_state_change=tracker.on_stanza_state_change
         )
@@ -1015,7 +1015,7 @@ class MUCClient(aioxmpp.service.Service):
         presence.xep0045_muc = muc_xso.GenericExt()
         presence.xep0045_muc.password = password
         presence.xep0045_muc.history = history
-        self.client.stream.enqueue_stanza(presence)
+        self.client.stream.enqueue(presence)
 
     def _stream_established(self):
         self.logger.debug("stream established, (re-)connecting to %d mucs",
@@ -1084,7 +1084,7 @@ class MUCClient(aioxmpp.service.Service):
                 type_=aioxmpp.structs.PresenceType.UNAVAILABLE,
             )
             unjoin.xep0045_muc = muc_xso.GenericExt()
-            self.client.stream.enqueue_stanza(unjoin)
+            self.client.stream.enqueue(unjoin)
 
     def _pending_on_enter(self, presence, occupant, **kwargs):
         mucjid = presence.from_.bare()